Move fetch batches out
