+++ expertise/models/specter2_scincl/specter.py
@@ -72,13 +72,6 @@ class Specter2Predictor(Predictor):
         self.model.to(self.cuda_device)
         self.model.eval()
 
-    def _fetch_batches(self, dict_data, batch_size):
-        iterator = iter(dict_data.items())
-        for _ in itertools.count():
-            batch = list(itertools.islice(iterator, batch_size))
-            if not batch:
-                break
-            yield batch
 
     def _batch_predict(self, batch_data):
         jsonl_out = []
